+++ backend/src/constants/index.ts
@@ -37,6 +37,8 @@ const SocketEventEnum = Object.freeze({
   USER_KICK_FROM_LIVE:'user_kick_from_live',
   BAN_VIEWER_FROM_MY_LIVE_STREAM:'ban_user_from_my_live_stream',
   NOTIFY_USER:'notify_user',
+  FOLLOW_USER:'follow_user',
+  UNFOLLOW_USER:'unfollow_user',
 
 });
